requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.1
django-ckeditor~=6.1.0asgiref==3.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-render-partial==0.4
django-widget-tweaks==1.4.8
Pillow==8.3.2
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
